requirements.txt
/app/MU-LLaMA/wheels/redis_layer-0.1.0-py3-none-any.whl
redis==5.2.1
fairscale==0.4.13
sentencepiece==0.1.99
Pillow==9.5.0
opencv-python==4.7.0.72
gradio==3.45.0
tqdm==4.65.0
autofaiss==2.15.8
tensorboard==2.12.3
diffusers==0.17.0
plotly==5.15.0
easydict==1.10
https://github.com/unlimblue/KNN_CUDA/releases/download/0.2/KNN_CUDA-0.2-py3-none-any.whl
huggingface_hub==0.15.1
transformers==4.30.2
python-dotenv
pydub
pytorchvideo
openai
librosa
requests
soundfile
scipy